Apply review fixes
