+++ utbot-junit-contest/src/main/kotlin/org/utbot/contest/ContestEstimator.kt
@@ -427,7 +427,7 @@ fun main(args: Array<String>) {
     // very special case when you run your project directly from IntellijIDEA omitting command line arguments
     if (args.isEmpty()) {
         processedClassesThreshold = 9999 //change to change number of classes to run
-        val timeLimit = 20 // increase if you want to debug something
+        val timeLimit = 30 // increase if you want to debug something
         val fuzzingRatio = 0.1 // sets fuzzing ratio to total test generation
 
         // Uncomment it for debug purposes:
@@ -447,8 +447,8 @@ fun main(args: Array<String>) {
 //        tools = listOf(Tool.EvoSuite)
 
         // config for SBST 2022
-        methodFilter = null
-        projectFilter = listOf("fastjson-1.2.50", "guava-26.0", "seata-core-0.5.0", "spoon-core-7.0.0")
+        methodFilter = "com.alibaba.fastjson.asm.ByteVector.*"
+        projectFilter = listOf("fastjson-1.2.50")
         tools = listOf(mainTool)
 
         estimatorArgs = arrayOf(
